sarvam-analyzer/requirements.txt
flask
flask-cors
python-dotenv
pdfplumber
requests
pytesseract
pillow
spacy
pydantic
pdf2image
